// apps/chataigne/ui/src/lib/systems/state_machine/generated/ManagedRegionDefinitionDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FormulaSurfaceItemKindDto } from "./FormulaSurfaceItemKindDto";
import type { ManagedFilterValueModeDto } from "./ManagedFilterValueModeDto";
import type { ManagedRegionKindDto } from "./ManagedRegionKindDto";
import type { ManagedSocketRefDto } from "./ManagedSocketRefDto";

export type ManagedRegionDefinitionDto = { id: string, kind: ManagedRegionKindDto, label: string, input_socket: ManagedSocketRefDto | null, output_socket: ManagedSocketRefDto | null, accepted_roles: Array<FormulaSurfaceItemKindDto>, filter_value_mode: ManagedFilterValueModeDto, };