// apps/chataigne/ui/src/lib/systems/state_machine/generated/MappingOutputSourceDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MappingConstantDto } from "./MappingConstantDto";
import type { MappingValueComponentDto } from "./MappingValueComponentDto";

export type MappingOutputSourceDto = { "kind": "whole" } | { "kind": "element", id: string, } | { "kind": "component", element: string | null, component: MappingValueComponentDto, } | { "kind": "constant", value: MappingConstantDto, };