// apps/chataigne/ui/src/lib/systems/state_machine/generated/MappingValueShapeDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MappingShapeElementDto } from "./MappingShapeElementDto";
import type { MappingShapeKindDto } from "./MappingShapeKindDto";

export type MappingValueShapeDto = { kind: MappingShapeKindDto, elements: Array<MappingShapeElementDto>, };